feat: reimplementing cli tests
