+++ tomb-common/src/types/blockstore/car/carv1/carv1blockstore.rs
@@ -4,7 +4,7 @@ use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use std::{
     borrow::Cow,
-    fs::{remove_file, rename, File, OpenOptions},
+    fs::{File, OpenOptions},
     io::{Seek, SeekFrom},
     path::{Path, PathBuf},
 };
