+++ tomb-common/src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod config;
-pub mod disk;
 pub mod serialize;
 pub mod tests;
